feat: Move streamlit code to its folder
